+++ NEWS.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.9] - 2024-12-20
+## [0.10.10] - 2024-12-20
+
+### Added
+
+* the `Segre` manifold
+* the `WarpedMetric` for the `Segre`manifold
+
+## [0.10.9] - 2024-12-16
 
 ### Added